modulize per function, per object(gyobutton.js -> gyo.js, style.js, button.js, togglebutton.js)
