+++ index.html
@@ -6,7 +6,9 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>gyobutton example</title>
     <link rel="stylesheet" type="text/css" href="./src/css/index.css">
+    <!-- <script type="text/javascript" src="./dist/js/gyobutton.js"></script> -->
     <script type="text/javascript" src="./dist/js/gyobutton.js"></script>
+
 </head>
 <body>
     <button id="i-1" class="button">BUTTON</button>
