+++ src/js/gyo.js
@@ -1,0 +1,8 @@
+import GyoButton from './button';
+import GyoToggleButton from './togglebutton';
+
+let Gyo 
+window.Gyo = Gyo = {};
+
+Gyo.button = GyoButton;
+Gyo.toggleButton = GyoToggleButton;
