+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-    entry: ['@babel/polyfill', './src/js/gyobutton.js'/*, './src/css/index.scss'*/],
+    entry: ['@babel/polyfill', './src/js/gyo.js'/*, './src/css/index.scss'*/],
 
     output: {
         path: path.resolve(__dirname, 'dist/js'),
